Adding some stub data

+++ app/src/main/java/com/example/homiez/objects/Posting.java
@@ -7,14 +7,14 @@ public class Posting {
     private String postingId;
     private String title;  //post id.
     private User user;
-    private int price;  //we could set a price, and the budget is around that price, for instamce around 500$
+    private double price;  //we could set a price, and the budget is around that price, for instamce around 500$
     private String location;
     private String type;
     private String description;
     private ArrayList<User>attachedUsers;
 
 
-    public Posting(String postingId,String title,User user, int price, String location, String type, String description){
+    public Posting(String postingId,String title,User user, double price, String location, String type, String description){
         this.postingId=postingId;
         this.title=title;
         this.user=user;
@@ -47,7 +47,7 @@ public class Posting {
     public void setPrice(int price){
         this.price=price;
     }
-    public int getPrice(){
+    public double getPrice(){
         return this.price;
     }
